+++ src/server/api/routers/adminRoute.ts
@@ -20,12 +20,7 @@ import axios from "axios";
 import { updateLocalConf } from "~/utils/planet";
 import jwt from "jsonwebtoken";
 import { networkRouter } from "./networkRouter";
-import {
-	API_TOKEN_SECRET,
-	decrypt,
-	encrypt,
-	generateInstanceSecret,
-} from "~/utils/encryption";
+import { decrypt, encrypt, generateInstanceSecret } from "~/utils/encryption";
 import { SMTP_SECRET } from "~/utils/encryption";
 import { ZT_FOLDER } from "~/utils/ztApi";
 import { isRunningInDocker } from "~/utils/docker";
@@ -1050,48 +1045,4 @@ export const adminRouter = createTRPCRouter({
 			}
 		}
 	}),
-	getApiToken: adminRoleProtectedRoute.query(async ({ ctx }) => {
-		return await ctx.prisma.aPIToken.findMany({
-			where: {
-				userId: ctx.session.user.id,
-			},
-		});
-	}),
-	addApiToken: adminRoleProtectedRoute
-		.input(
-			z.object({
-				name: z.string().min(5).max(50),
-			}),
-		)
-		.mutation(async ({ ctx, input }) => {
-			const token_content: string = JSON.stringify({
-				name: input.name,
-				userId: ctx.session.user.id,
-			});
-
-			const token_hash = encrypt(token_content, generateInstanceSecret(API_TOKEN_SECRET));
-			const token = await ctx.prisma.aPIToken.create({
-				data: {
-					token: token_hash,
-					name: input.name,
-					userId: ctx.session.user.id,
-				},
-			});
-			return token;
-		}),
-
-	deleteApiToken: adminRoleProtectedRoute
-		.input(
-			z.object({
-				id: z.number(),
-			}),
-		)
-		.mutation(async ({ ctx, input }) => {
-			return await ctx.prisma.aPIToken.delete({
-				where: {
-					id: input.id,
-					userId: ctx.session.user.id,
-				},
-			});
-		}),
 });
